Serviço de Pessoas - tratamento de exceções: adicionando exeções no FornecedorService

+++ pessoa-service/src/main/java/com/ms/rr/pessoa_service/domain/service/FornecedorServiceImpl.java
@@ -3,6 +3,8 @@ package com.ms.rr.pessoa_service.domain.service;
 import com.ms.rr.pessoa_service.domain.dto.in.CreateFornecedor;
 import com.ms.rr.pessoa_service.domain.dto.in.UpdateFornecedor;
 import com.ms.rr.pessoa_service.domain.dto.out.FornecedorResponse;
+import com.ms.rr.pessoa_service.domain.exception.BaseErrorMessage;
+import com.ms.rr.pessoa_service.domain.exception.FornecedorException;
 import com.ms.rr.pessoa_service.domain.port.input.FornecedorUseCase;
 import com.ms.rr.pessoa_service.domain.port.output.FornecedorOutputPort;
 import com.ms.rr.pessoa_service.domain.port.output.PessoaCriadaOutpuPort;
@@ -28,6 +30,14 @@ public class FornecedorServiceImpl implements FornecedorUseCase {
     @Transactional
     @Override
     public void salvar(CreateFornecedor createFornecedor) {
+        if (fornecedorOutputPort.emailExists(createFornecedor.email())) {
+            throw new FornecedorException(BaseErrorMessage.EXISTS_PESSOA_EMAIL.getMessage());
+        }
+
+        if (fornecedorOutputPort.cnpjExists(createFornecedor.cnpj())) {
+            throw new FornecedorException(BaseErrorMessage.EXISTS_PESSOA_FORNECEDOR_CNPJ.getMessage());
+        }
+
         fornecedorOutputPort.save(createFornecedor.toDomain());
         pessoaCriadaOutpuPort.sendMessage(eventFromFornecedorDomain(createFornecedor.toDomain()));
     }
@@ -36,7 +46,7 @@ public class FornecedorServiceImpl implements FornecedorUseCase {
     public FornecedorResponse buscarPorId(Long id) {
         return fornecedorOutputPort.findById(id)
                 .map(FornecedorResponse::fromDomain)
-                .orElseThrow();
+                .orElseThrow(() -> new FornecedorException(BaseErrorMessage.FORNECEDOR_NOT_FOUND.getMessage()));
     }
 
     @Override
@@ -49,12 +59,18 @@ public class FornecedorServiceImpl implements FornecedorUseCase {
     @Transactional
     @Override
     public void atualizar(Long id, UpdateFornecedor updateFornecedor) {
+        if (fornecedorOutputPort.findById(id).isEmpty()) {
+            throw new FornecedorException(BaseErrorMessage.FORNECEDOR_NOT_FOUND.getMessage());
+        }
         fornecedorOutputPort.update(id, updateFornecedor.toDomain(id));
     }
 
     @Transactional
     @Override
     public void excluir(Long id) {
+        if (fornecedorOutputPort.findById(id).isEmpty()) {
+            throw new FornecedorException(BaseErrorMessage.FORNECEDOR_NOT_FOUND.getMessage());
+        }
         fornecedorOutputPort.deleteById(id);
     }
 
@@ -70,7 +86,11 @@ public class FornecedorServiceImpl implements FornecedorUseCase {
 
     @Override
     public FornecedorResponse buscarPorCnpj(String cnpj) {
+        FornecedorDomain fornecedor = fornecedorOutputPort.findFornecedorByCnpj(cnpj);
+        if (fornecedor == null) {
+            throw new FornecedorException(BaseErrorMessage.FORNECEDOR_NOT_FOUND.getMessage());
+        }
         return FornecedorResponse
-                .fromDomain(fornecedorOutputPort.findFornecedorByCnpj(cnpj));
+                .fromDomain(fornecedor);
     }
 }
